UI and js function update
Completed all JavaScript buttons onclick functions that need to display the property value of list-item to the recipe ingredients and its instructions needed to make any cake required.

Last attention to CSS paragraph width  will be updated inside the media queries for mobile use towards the width that is set to exactly 400px which needs another 4h hours to test and implement for CSS property value to replace 400px with auto or smaller fixed sizes.